Regions weg, database omgehangen, repository tussen context en controller (of ander object)
