+++ ms18-applicatie.sln
@@ -7,7 +7,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ms18-applicatie", "ms18-app
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ms18-database", "ms18-database\ms18-database.csproj", "{4C7B4B5F-651F-47DE-9EA3-E827F8965208}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ms18-applicatie-console", "ms18-applicatie-console\ms18-applicatie-console.csproj", "{88284C64-629D-4389-8727-4E13BDF5413A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ms18-applicatie-console", "ms18-applicatie-console\ms18-applicatie-console.csproj", "{88284C64-629D-4389-8727-4E13BDF5413A}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ms18-sharedKernel", "ms18-sharedKernel\ms18-sharedKernel.csproj", "{D4834E1E-AFCC-445B-8D75-AEC14E10B40C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,6 +29,10 @@ Global
 		{88284C64-629D-4389-8727-4E13BDF5413A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{88284C64-629D-4389-8727-4E13BDF5413A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{88284C64-629D-4389-8727-4E13BDF5413A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D4834E1E-AFCC-445B-8D75-AEC14E10B40C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D4834E1E-AFCC-445B-8D75-AEC14E10B40C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D4834E1E-AFCC-445B-8D75-AEC14E10B40C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D4834E1E-AFCC-445B-8D75-AEC14E10B40C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
