+++ ms18-database/teamC/AdminContext/IMemberRepository.cs
@@ -1,0 +1,10 @@
+﻿
+namespace Maasgroep.Database.Members
+{
+    public interface IMemberRepository
+    {
+        void AddMember(); // Member kan eventueel nog weg, gewoon Add()
+        void RemoveMember(); 
+
+    }
+}
