+++ ms18-database/teamC/PhotoContext/IPhotoRepository.cs
@@ -1,0 +1,9 @@
+﻿
+namespace Maasgroep.Database.Photos
+{
+    public interface IPhotoRepository
+    {
+        void AddPhoto();
+        void RemovePhoto();
+    }
+}
