+++ ms18-database/teamC/ReceiptContext/IReceiptRepository.cs
@@ -1,0 +1,12 @@
+﻿
+namespace Maasgroep.Database.Receipts
+{
+    public interface IReceiptRepository
+    {
+        void AddReceipt();
+
+        void ModifyReceipt();
+
+        // Nadenken over GetReceipt en model
+    }
+}
