+++ ms18-database/teamC/ReceiptContext/Tables/Receipt.cs
@@ -1,7 +1,8 @@
-﻿
-namespace Maasgroep.Database
+﻿using Maasgroep.Database.Members;
+
+namespace Maasgroep.Database.Receipts
 {
-	public record Receipt
+    internal record Receipt
 	{
 		public long Id { get; set; }
 		public decimal? Amount { get; set; }
