+++ ms18-database/teamC/ReceiptContext/Tables/ReceiptStatus.cs
@@ -1,7 +1,8 @@
-﻿
-namespace Maasgroep.Database
+﻿using Maasgroep.Database.Members;
+
+namespace Maasgroep.Database.Receipts
 {
-	public record ReceiptStatus
+    internal record ReceiptStatus
 	{
 		public long Id { get; set; }
 		public string Name { get; set; }
