+++ ms18-sharedKernel/teamC/Receipts/IReceipt.cs
@@ -1,0 +1,7 @@
+﻿
+namespace Maasgroep.SharedKernel.Receipts
+{
+    internal interface IReceipt
+    {
+    }
+}
